core/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
click==8.3.3
ecdsa==0.19.2
fastapi==0.136.1
greenlet==3.4.0
h11==0.16.0
idna==3.13
Mako==1.3.11
MarkupSafe==3.0.3
psycopg2-binary==2.9.12
pyasn1==0.6.3
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
python-dotenv==1.2.2
python-jose==3.5.0
redis==7.4.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
